+++ pystac_client/conformance.py
@@ -1,4 +1,4 @@
-"""This module defines a :class:`ConformanceClass` class that can be used to check conformance to a spec based on the
+"""This module defines a :class:`ConformanceClasses` class that can be used to check conformance to a spec based on the
 presence of a conformance URI. Each instance has a single :attr:`ConformanceClass.uri` property that represents the
 most up-to-date official conformance URI for that spec. Because the STAC API spec has been rapidly evolving, many APIs
 publish other conformance URIs that were defined in previous iterations of the spec, or are defined by other entities.
@@ -29,18 +29,22 @@ True
 """
 from enum import Enum
 
-STAC_PREFIXES = ['https://api.stacspec.org/v1.0.0-beta.2', 'https://api.stacspec.org/v1.0.0-beta.1']
+import re
 
 
 class ConformanceClasses(Enum):
-    CORE = [f"{p}/core" for p in STAC_PREFIXES]
-    ITEM_SEARCH = [f"{p}/item-search" for p in STAC_PREFIXES]
-    CONTEXT = [f"{p}/item-search#context" for p in STAC_PREFIXES]
-    FIELDS = [f"{p}/item-search#fields" for p in STAC_PREFIXES]
-    SORT = [f"{p}/item-search#sort" for p in STAC_PREFIXES]
-    QUERY = [f"{p}/item-search#query" for p in STAC_PREFIXES]
-    FILTER = [f"{p}/item-search#filter" for p in STAC_PREFIXES]
-    COLLECTIONS = ['http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30']
+
+    stac_prefix = re.escape("https://api.stacspec.org/v1.0.")
+
+    # defined conformance classes regexes
+    CORE = fr"{stac_prefix}(.*){re.escape('/core')}"
+    ITEM_SEARCH = fr"{stac_prefix}(.*){re.escape('/item-search')}"
+    CONTEXT = fr"{stac_prefix}(.*){re.escape('/item-search#context')}"
+    FIELDS = fr"{stac_prefix}(.*){re.escape('/item-search#fields')}"
+    SORT = fr"{stac_prefix}(.*){re.escape('/item-search#sort')}"
+    QUERY = fr"{stac_prefix}(.*){re.escape('/item-search#query')}"
+    FILTER = fr"{stac_prefix}(.*){re.escape('/item-search#filter')}"
+    COLLECTIONS = re.escape("http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30")
 
 
 CONFORMANCE_URIS = {c.name: c.value for c in ConformanceClasses}